requirements.txt
streamlit
pyyaml
requests

